hehaokun1@huawei.com

+++ frameworks/innerkitsimpl/converter/src/post_proc.cpp
@@ -980,7 +980,7 @@ static bool PixelMapPostProcWithGL(PixelMap &sourcePixelMap, GPUTransformData &t
     if (dmaMode.isSourceDma && !dmaMode.isTargetDma) {
         IMAGE_LOGI("slr_gpu PixelMapPostProcWithGL no-padding DMA does not support GPU writeback, fallback output");
     }
-    size_t buffersize = static_cast<uint64_t>(desiredSize.width) * desiredSize.height * 4; // 4: 4 bytes per pixel
+    size_t buffersize = static_cast<uint64_t>(desiredSize.width) * static_cast<uint64_t>(desiredSize.height) * 4;
     MemoryData memoryData = {nullptr, buffersize, "PixelMapPostProcWithGL", desiredSize};
     memoryData.usage = noPaddingUsage;
     std::unique_ptr<AbsMemory> dstMemory = MemoryManager::CreateMemory(dmaMode.outputAllocType, memoryData);
@@ -990,7 +990,8 @@ static bool PixelMapPostProcWithGL(PixelMap &sourcePixelMap, GPUTransformData &t
     if (dmaMode.isTargetDma) {
         SurfaceBuffer* sbBuffer = reinterpret_cast<SurfaceBuffer*>(dstMemory->extend.data);
         outputStride = sbBuffer->GetStride();
-        buffersize = static_cast<uint32_t>(static_cast<uint64_t>(sbBuffer->GetStride()) * desiredSize.height);
+        buffersize = static_cast<uint32_t>(static_cast<uint64_t>(sbBuffer->GetStride()) *
+            static_cast<uint64_t>(desiredSize.height));
     }
     PixelMapProgramManager::BuildShader();
     bool ret = true;
@@ -1016,7 +1017,7 @@ static bool PixelMapPostProcWithGL(PixelMap &sourcePixelMap, GPUTransformData &t
         return false;
     }
     sourcePixelMap.SetPixelsAddr(dstMemory->data.data, dstMemory->extend.data,
-        static_cast<uint64_t>(desiredSize.height) * outputStride, dmaMode.outputAllocType, nullptr);
+        static_cast<uint64_t>(desiredSize.height) * static_cast<uint64_t>(outputStride), dmaMode.outputAllocType, nullptr);
     ImageInfo info;
     info.size = desiredSize;
     info.pixelFormat = PixelFormat::RGBA_8888;
